requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
databricks-sql-connector==2.8.0
pandas==2.2.2
numpy==1.26.4
